Module pallet_xcm_bridge::congestion

The module contains utilities for handling congestion between the bridge hub and routers.

Structs

BlobDispatcherWithChannelStatus

Adapter that ties together the [DispatchBlob] trait with the [DispatchChannelStatusProvider] trait. The idea is that [DispatchBlob] triggers message dispatch/delivery on the receiver side, while [DispatchChannelStatusProvider] provides a status check to ensure the dispatch channel is active (not congested).

CongestionLimits

Limits for handling congestion.

HereOrLocalConsensusXcmChannelManager

Switches the implementation of [LocalXcmChannelManager] based on the local_origin.

UpdateBridgeStatusXcmChannelManager

Manages the local XCM channels by sending XCM messages with the update_bridge_status extrinsic to the local_origin. The XcmProvider type converts the encoded call to XCM, which is then sent by XcmSender to the local_origin. This is useful, for example, when a router with [xcm::prelude::ExportMessage] is deployed on a different chain, and we want to control congestion by sending XCMs.