Function pallet_xcm_bridge::pallet::dispatchables::open_bridge

pub fn open_bridge<T: Config<I>, I: 'static>(
    bridge_destination_universal_location: Box<VersionedInteriorLocation>,
    maybe_notify: Option<Receiver>
)

Open a bridge between two locations.

The caller must be within the T::OpenBridgeOrigin filter (presumably: a sibling parachain or a parent relay chain). The bridge_destination_universal_location must be a destination within the consensus of the T::BridgedNetwork network.

The BridgeDeposit amount is reserved on the caller account. This deposit is unreserved after bridge is closed.

The states after this call: bridge is Opened, outbound lane is Opened, inbound lane is Opened.

Optional maybe_notify holds data about the bridge_origin_relative_location where notifications can be sent to handle congestion. The notification contains an encoded tuple (BridgeId, bool), where bool is the is_congested flag.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::open_bridge.