Trait pallet_xcm_bridge::pallet::Config

pub trait Config<I: 'static = ()>: BridgeMessagesConfig<Self::BridgeMessagesPalletInstance> {

    type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>;
    type WeightInfo: WeightInfo;
    type UniversalLocation: Get<InteriorLocation>;
    type BridgedNetwork: Get<Location>;
    type BridgeMessagesPalletInstance: 'static;
    type MessageExportPrice: Get<Assets>;
    type DestinationVersion: GetVersion;
    type ForceOrigin: EnsureOrigin<<Self as SystemConfig>::RuntimeOrigin>;
    type OpenBridgeOrigin: EnsureOrigin<<Self as SystemConfig>::RuntimeOrigin, Success = Location>;
    type BridgeOriginAccountIdConverter: ConvertLocation<AccountIdOf<ThisChainOf<Self, I>>>;
    type BridgeDeposit: Get<BalanceOf<ThisChainOf<Self, I>>>;
    type Currency: MutateHold<AccountIdOf<ThisChainOf<Self, I>>, Balance = BalanceOf<ThisChainOf<Self, I>>, Reason = Self::RuntimeHoldReason>;
    type RuntimeHoldReason: From<HoldReason<I>>;
    type AllowWithoutBridgeDeposit: Contains<Location>;
    type LocalXcmChannelManager: LocalXcmChannelManager<BridgeId>;
    type BlobDispatcher: DispatchBlob + DispatchChannelStatusProvider;
    type CongestionLimits: Get<CongestionLimits>;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type WeightInfo: WeightInfo

Benchmarks results from runtime we’re plugged into.

type UniversalLocation: Get<InteriorLocation>

Runtime’s universal location.

type BridgedNetwork: Get<Location>

Bridged network as relative location of bridged GlobalConsensus.

type BridgeMessagesPalletInstance: 'static

Associated messages pallet instance that bridges us with the BridgedNetworkId consensus.

type MessageExportPrice: Get<Assets>

Price of single message export to the bridged consensus (Self::BridgedNetwork).

type DestinationVersion: GetVersion

Checks the XCM version for the destination.

type ForceOrigin: EnsureOrigin<<Self as SystemConfig>::RuntimeOrigin>

The origin that is allowed to call privileged operations on the pallet, e.g. open/close bridge for locations.

type OpenBridgeOrigin: EnsureOrigin<<Self as SystemConfig>::RuntimeOrigin, Success = Location>

A set of XCM locations within local consensus system that are allowed to open bridges with remote destinations.

type BridgeOriginAccountIdConverter: ConvertLocation<AccountIdOf<ThisChainOf<Self, I>>>

A converter between a location and a sovereign account.

type BridgeDeposit: Get<BalanceOf<ThisChainOf<Self, I>>>

Amount of this chain native tokens that is reserved on the sibling parachain account when bridge open request is registered.

type Currency: MutateHold<AccountIdOf<ThisChainOf<Self, I>>, Balance = BalanceOf<ThisChainOf<Self, I>>, Reason = Self::RuntimeHoldReason>

Currency used to pay for bridge registration.

type RuntimeHoldReason: From<HoldReason<I>>

The overarching runtime hold reason.

type AllowWithoutBridgeDeposit: Contains<Location>

Do not hold Self::BridgeDeposit for the location of Self::OpenBridgeOrigin. For example, it is possible to make an exception for a system parachain or relay.

type LocalXcmChannelManager: LocalXcmChannelManager<BridgeId>

Local XCM channel manager. Dedicated to exporting capabilities when handling congestion with the sending side.

type BlobDispatcher: DispatchBlob + DispatchChannelStatusProvider

XCM-level dispatcher for inbound bridge messages.

type CongestionLimits: Get<CongestionLimits>

Limits for handling congestion.

Object Safety

This trait is not object safe.